wgpu: Separated srgb out of Surface
