+++ render/wgpu/src/frame.rs
@@ -28,16 +28,6 @@ impl<'a> Frame<'a> {
         }
     }
 
-    pub fn prep_srgb_copy(
-        &mut self,
-        bind_group: &'a wgpu::BindGroup,
-        pipeline: &'a wgpu::RenderPipeline,
-    ) {
-        self.render_pass.set_pipeline(&pipeline);
-
-        self.render_pass.set_bind_group(2, bind_group, &[]);
-    }
-
     pub fn finish(self) {
         self.uniform_buffers.finish()
     }
